Modified int types to follow new isize/usize convention.
uints which were used as pointers are now usize, all others are now i32/u32.

+++ arch/x86/cpu/gdt.rs
@@ -6,7 +6,7 @@
 
 use core::marker::Copy;
 
-const GDT_COUNT: uint = 5;
+const GDT_COUNT: usize = 5;
 static mut gdt_entries: [GDTEntry; GDT_COUNT] = [GDTEntry { limit_low: 0, base_low: 0, base_middle: 0, access: 0, granularity: 0, base_high: 0 }; GDT_COUNT];
 static mut gdt_ptr: GDTPointer = GDTPointer { limit: 0, base: 0 };
 
@@ -27,7 +27,7 @@ impl Copy for GDTEntry {}
 struct GDTPointer
 {
 	limit: u16,
-	base: u32
+	base: usize
 }
 
 pub fn init_gdt()
@@ -35,7 +35,7 @@ pub fn init_gdt()
 	unsafe
 	{
 		gdt_ptr.limit = (::core::mem::size_of::<GDTEntry>() * GDT_COUNT - 1) as u16;
-		gdt_ptr.base = &gdt_entries as *const [GDTEntry; GDT_COUNT] as u32;
+		gdt_ptr.base = &gdt_entries as *const [GDTEntry; GDT_COUNT] as usize;
 
 		gdt_set_gate(0, 0, 0, 0, 0);
 		gdt_set_gate(1, 0, 0xFFFFFFFF, 0x9A, 0xCF);
@@ -43,11 +43,11 @@ pub fn init_gdt()
 		gdt_set_gate(3, 0, 0xFFFFFFFF, 0xFA, 0xCF);
 		gdt_set_gate(4, 0, 0xFFFFFFFF, 0xF2, 0xCF);
 
-		gdt_flush(&gdt_ptr as *const GDTPointer as u32);
+		gdt_flush(&gdt_ptr as *const GDTPointer as usize);
 	};
 }
 
-unsafe fn gdt_set_gate(n: uint, base: u32, limit: u32, access: u8, gran: u8)
+unsafe fn gdt_set_gate(n: usize, base: usize, limit: usize, access: u8, gran: u8)
 {
 	gdt_entries[n].base_low = (base & 0xFFFF) as u16;
 	gdt_entries[n].base_middle = ((base >> 16) & 0xFF) as u8;
@@ -62,5 +62,5 @@ unsafe fn gdt_set_gate(n: uint, base: u32, limit: u32, access: u8, gran: u8)
 
 extern
 {
-	fn gdt_flush(pointer: u32);
+	fn gdt_flush(pointer: usize);
 }
